+++ pkg/vm/engine/disttae/pk_filter_mem.go
@@ -27,34 +27,36 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/disttae/logtailreplay"
 )
 
-type memPKFilter struct {
-	op             int
-	packed         [][]byte
-	isVec          bool
-	IsValid        bool
-	Iter           logtailreplay.RowsIter
-	delIterFactory func(blkId types.Blockid) logtailreplay.RowsIter
+type MemPKFilter struct {
+	op      int
+	packed  [][]byte
+	isVec   bool
+	isValid bool
+	TS      types.TS
+
+	SpecFactory func(f *MemPKFilter) logtailreplay.PrimaryKeyMatchSpec
 }
 
-func newMemPKFilter(
+func newMemPKFilterInProgress(
 	tableDef *plan.TableDef,
 	ts timestamp.Timestamp,
-	state *logtailreplay.PartitionState,
 	packerPool *fileservice.Pool[*types.Packer],
 	basePKFilter basePKFilter,
-) (filter memPKFilter) {
-	defer func() {
-		if filter.Iter == nil {
-			filter.IsValid = true
-			filter.Iter = state.NewRowsIter(
-				types.TimestampToTS(ts),
-				nil,
-				false,
-			)
-		}
-	}()
+) (filter MemPKFilter) {
+	//defer func() {
+	//	if filter.iter == nil {
+	//		filter.isValid = true
+	//		filter.iter = state.NewRowsIter(
+	//			types.TimestampToTS(ts),
+	//			nil,
+	//			false,
+	//		)
+	//	}
+	//}()
 
-	if tableDef.Pkey == nil || !basePKFilter.valid {
+	filter.TS = types.TimestampToTS(ts)
+
+	if !basePKFilter.valid || tableDef.Pkey == nil {
 		return
 	}
 
@@ -211,47 +213,40 @@ func newMemPKFilter(
 		return
 	}
 
-	filter.tryConstructPrimaryKeyIndexIter(ts, state)
+	filter.tryConstructPrimaryKeyIndexIterInProgress(ts)
 	return
 }
 
-func (f *memPKFilter) String() string {
+func (f *MemPKFilter) String() string {
 	var buf bytes.Buffer
 	buf.WriteString(
 		fmt.Sprintf("InMemPKFilter{op: %d, isVec: %v, isValid: %v, val: %v, data(len=%d)",
-			f.op, f.isVec, f.IsValid, f.packed, len(f.packed),
+			f.op, f.isVec, f.isValid, f.packed, len(f.packed),
 		))
 	return buf.String()
 }
 
-func (f *memPKFilter) SetNull() {
-	f.IsValid = false
+func (f *MemPKFilter) SetNull() {
+	f.isValid = false
 }
 
-func (f *memPKFilter) SetFullData(op int, isVec bool, val ...[]byte) {
+func (f *MemPKFilter) SetFullData(op int, isVec bool, val ...[]byte) {
 	f.packed = append(f.packed, val...)
 	f.op = op
 	f.isVec = isVec
-	f.IsValid = true
+	f.isValid = true
 }
 
-func (f *memPKFilter) tryConstructPrimaryKeyIndexIter(
-	ts timestamp.Timestamp,
-	state *logtailreplay.PartitionState) {
-	if !f.IsValid {
+func (f *MemPKFilter) tryConstructPrimaryKeyIndexIterInProgress(ts timestamp.Timestamp) {
+	if !f.isValid {
 		return
 	}
 
 	switch f.op {
 	case function.EQUAL, function.PREFIX_EQ:
-		f.Iter = state.NewPrimaryKeyIter(
-			types.TimestampToTS(ts),
-			logtailreplay.Prefix(f.packed[0]),
-		)
-		f.delIterFactory = func(blkId types.Blockid) logtailreplay.RowsIter {
-			return state.NewPrimaryKeyDelIter(
-				types.TimestampToTS(ts),
-				logtailreplay.Prefix(f.packed[0]), blkId)
+		//spec = logtailreplay.Prefix(f.packed[0])
+		f.SpecFactory = func(f *MemPKFilter) logtailreplay.PrimaryKeyMatchSpec {
+			return logtailreplay.Prefix(f.packed[0])
 		}
 
 	case function.IN, function.PREFIX_IN:
@@ -259,37 +254,21 @@ func (f *memPKFilter) tryConstructPrimaryKeyIndexIter(
 		if len(f.packed) > 128 {
 			return
 		}
-
-		f.Iter = state.NewPrimaryKeyIter(
-			types.TimestampToTS(ts),
-			logtailreplay.InKind(f.packed, f.op),
-		)
-		f.delIterFactory = func(blkId types.Blockid) logtailreplay.RowsIter {
-			return state.NewPrimaryKeyDelIter(
-				types.TimestampToTS(ts),
-				logtailreplay.InKind(f.packed, f.op), blkId)
+		//spec = logtailreplay.InKind(f.packed, f.op)
+		f.SpecFactory = func(f *MemPKFilter) logtailreplay.PrimaryKeyMatchSpec {
+			return logtailreplay.InKind(f.packed, f.op)
 		}
 
 	case function.LESS_EQUAL, function.LESS_THAN:
-		f.Iter = state.NewPrimaryKeyIter(
-			types.TimestampToTS(ts),
-			logtailreplay.LessKind(f.packed[0], f.op == function.LESS_EQUAL),
-		)
-		f.delIterFactory = func(blkId types.Blockid) logtailreplay.RowsIter {
-			return state.NewPrimaryKeyDelIter(
-				types.TimestampToTS(ts),
-				logtailreplay.LessKind(f.packed[0], f.op == function.LESS_EQUAL), blkId)
+		//spec = logtailreplay.LessKind(f.packed[0], f.op == function.LESS_EQUAL)
+		f.SpecFactory = func(f *MemPKFilter) logtailreplay.PrimaryKeyMatchSpec {
+			return logtailreplay.LessKind(f.packed[0], f.op == function.LESS_EQUAL)
 		}
 
 	case function.GREAT_EQUAL, function.GREAT_THAN:
-		f.Iter = state.NewPrimaryKeyIter(
-			types.TimestampToTS(ts),
-			logtailreplay.GreatKind(f.packed[0], f.op == function.GREAT_EQUAL),
-		)
-		f.delIterFactory = func(blkId types.Blockid) logtailreplay.RowsIter {
-			return state.NewPrimaryKeyDelIter(
-				types.TimestampToTS(ts),
-				logtailreplay.GreatKind(f.packed[0], f.op == function.GREAT_EQUAL), blkId)
+		//spec = logtailreplay.GreatKind(f.packed[0], f.op == function.GREAT_EQUAL)
+		f.SpecFactory = func(f *MemPKFilter) logtailreplay.PrimaryKeyMatchSpec {
+			return logtailreplay.GreatKind(f.packed[0], f.op == function.GREAT_EQUAL)
 		}
 
 	case function.BETWEEN, rangeLeftOpen, rangeRightOpen, rangeBothOpen, function.PREFIX_BETWEEN:
@@ -306,14 +285,11 @@ func (f *memPKFilter) tryConstructPrimaryKeyIndexIter(
 		case function.PREFIX_BETWEEN:
 			kind = 4
 		}
-		f.Iter = state.NewPrimaryKeyIter(
-			types.TimestampToTS(ts),
-			logtailreplay.BetweenKind(f.packed[0], f.packed[1], kind))
 
-		f.delIterFactory = func(blkId types.Blockid) logtailreplay.RowsIter {
-			return state.NewPrimaryKeyDelIter(
-				types.TimestampToTS(ts),
-				logtailreplay.BetweenKind(f.packed[0], f.packed[1], kind), blkId)
+		//spec = logtailreplay.BetweenKind(f.packed[0], f.packed[1], kind)
+		f.SpecFactory = func(f *MemPKFilter) logtailreplay.PrimaryKeyMatchSpec {
+			return logtailreplay.BetweenKind(f.packed[0], f.packed[1], kind)
 		}
 	}
+
 }
